FileUpload/requirements.txt
flask
pyodbc
mysql-connector-python
psycopg2-binary
